// src/app/core/api/models/app-schemas-acessibilidade-mapa-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AppSchemasAcessibilidadeMapaData } from '../models/app-schemas-acessibilidade-mapa-data';

/**
 * Resposta do mapa: descrição + lista de pontos georreferenciados.
 */
export interface AppSchemasAcessibilidadeMapaResponse {
  data: AppSchemasAcessibilidadeMapaData;

  /**
   * Identificador semântico do mapa.
   */
  descricao: string;
}