// src/app/core/api/models/filtros-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FiltrosData } from '../models/filtros-data';

/**
 * Resposta do endpoint genérico de filtros aplicáveis.
 */
export interface FiltrosResponse {
  data: FiltrosData;

  /**
   * Identificador semântico do recurso.
   */
  descricao: string;
}