// src/app/core/api/models/painel-escolas-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PainelEscolasData } from '../models/painel-escolas-data';

/**
 * Resposta paginada do gráfico de métricas por escola.
 */
export interface PainelEscolasResponse {
  data: PainelEscolasData;

  /**
   * Identificador semântico do recurso.
   */
  descricao: string;
}